requirements.txt
absl-py==0.8.1
astor==0.8.0
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
gym==0.15.3
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
opt-einsum==3.1.0
protobuf==3.10.0
pyglet==1.3.2
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
